LayoutTests/imported/w3c/web-platform-tests/workers/SharedWorker-detach-frame-in-error-event-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: SharedWorker


Harness Error (TIMEOUT), message = null

TIMEOUT Test frame detach in shared worker's error handler Test timed out
